Get agent self registration working
